requirements.txt
asgiref==3.7.2
Django==5.0
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
